refactor: CategoryName 페키지 이동
