+++ src/test/java/com/kustacks/kuring/worker/client/notice/KuisNoticeApiClientRetryTest.java
@@ -4,7 +4,7 @@ import com.kustacks.kuring.config.JsonConfig;
 import com.kustacks.kuring.config.RetryConfig;
 import com.kustacks.kuring.common.error.ErrorCode;
 import com.kustacks.kuring.common.error.InternalLogicException;
-import com.kustacks.kuring.worker.CategoryName;
+import com.kustacks.kuring.category.domain.CategoryName;
 import com.kustacks.kuring.common.utils.converter.KuisNoticeDTOToCommonFormatDTOConverter;
 import com.kustacks.kuring.worker.scrap.client.auth.KuisAuthManager;
 import com.kustacks.kuring.worker.scrap.client.notice.KuisNoticeApiClient;
